Remove AlertMacOS in favor of Alert
Fix merge issue

Get rid of workaround maybe?

Better type?

Fix codegen error

Remove AlertMacOS altogether

Space

Refactor test page logic

Delete AlertIOS

Make promptMacOS work

make example load

Fix Alert.prompt()

Move examples to AlertIOSExample

Remove changes to AlertExample

Remove more

fix error

More fixed
